net: Moved IANA defines to own header to avoid big include
